fix: failed psalm

+++ src/SDK/Trace/AutoRootSpan.php
@@ -44,8 +44,8 @@ class AutoRootSpan
         );
         $parent = Globals::propagator()->extract($request->getHeaders());
         $startTimeInNanos = array_key_exists('REQUEST_TIME_FLOAT', $request->getServerParams())
-            ? (int) ($request->getServerParams()['REQUEST_TIME_FLOAT'] * ClockInterface::NANOS_PER_SECOND)
-            : (int) (microtime(true) * ClockInterface::NANOS_PER_SECOND);
+            ? (int) ((float) $request->getServerParams()['REQUEST_TIME_FLOAT'] * (float) ClockInterface::NANOS_PER_SECOND)
+            : (int) (microtime(true) * (float) ClockInterface::NANOS_PER_SECOND);
         $span = $tracer->spanBuilder($request->getMethod())
             ->setSpanKind(SpanKind::KIND_SERVER)
             ->setStartTimestamp($startTimeInNanos)
